Fix MSVC compiler warning
reg. conversion from 'size_t' to 'int', possible loss of data (C4267)

+++ src/engraving/libmscore/edit.cpp
@@ -6066,7 +6066,7 @@ void Score::undoRemoveElement(EngravingItem* element)
     if (element->isMeasureRepeat()) {
         const MeasureRepeat* repeat = toMeasureRepeat(element);
         Measure* measure = repeat->firstMeasureOfGroup();
-        int staffIdx = repeat->staffIdx();
+        size_t staffIdx = repeat->staffIdx();
 
         for (int i = 0; i < repeat->numMeasures(); ++i) {
             undoChangeMeasureRepeatCount(measure, 0, staffIdx);
